vaccineapis/vaccineapp/requirements.txt
asgiref==3.8.1
Django==5.2
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
Markdown==3.8
mysqlclient==2.2.7
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
whitenoise==6.9.0
